Reorganized/renamed observable interfaces into own package.
Renamed all the interfaces to be consistent as I<ObservableType>Observer
because I've always had the design pattern be the last thing (like in
Controller,View, and Model).
