+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/IEndGameObservable.java
@@ -1,9 +1,9 @@
-package connect4Menu.model;
+package connect4Menu.model.observableinterfaces;
 
 import connect4Menu.model.player.Player;
 import connect4Menu.view.ObserverEndGame;
 
-public interface ObservableEndGame {
+public interface IEndGameObservable {
 
 	public void registerEndGameObserver(ObserverEndGame o);
