+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/IStartedTurnObservable.java
@@ -1,9 +1,9 @@
-package connect4Menu.model;
+package connect4Menu.model.observableinterfaces;
 
 import connect4Menu.model.player.Player;
 import connect4Menu.view.ObserverStartedTurn;
 
-public interface ObservableStartedTurn {
+public interface IStartedTurnObservable {
 
 	public void registerStartedTurnObserver(ObserverStartedTurn o);
